pyproject.toml
[tool.pdm]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Amanda de Oliveira Pires", email = "nanapires255@gmail.com"},
]
dependencies = [
    "django>=4.1.7",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.1.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}
